requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
requests
python-dotenv
